+++ apps/mrpt_apps_gui/robot-map-gui/gui/configWidget/CPointsConfig.cpp
@@ -35,14 +35,16 @@ void CPointsConfig::updateConfiguration(mrpt::maps::TMetricMapInitializer* optio
   mapDefination->genericMapParams.enableObservationLikelihood = m_ui->enableObservationLikelihood;
   mapDefination->genericMapParams.enableObservationInsertion = m_ui->enableObservationInsertion;
 
-  mapDefination->insertionOpts.minDistBetweenLaserPoints = m_ui->minDistBetweenLaserPoints->value();
+  mapDefination->insertionOpts.minDistBetweenLaserPoints =
+      static_cast<float>(m_ui->minDistBetweenLaserPoints->value());
   mapDefination->insertionOpts.addToExistingPointsMap = m_ui->addToExistingPointsMap->isChecked();
   mapDefination->insertionOpts.also_interpolate = m_ui->also_interpolate->isChecked();
   mapDefination->insertionOpts.fuseWithExisting = m_ui->fuseWithExisting->isChecked();
   mapDefination->insertionOpts.isPlanarMap = m_ui->isPlanarMap->isChecked();
-  mapDefination->insertionOpts.horizontalTolerance = m_ui->horizontalTolerance->value();
+  mapDefination->insertionOpts.horizontalTolerance =
+      static_cast<float>(m_ui->horizontalTolerance->value());
   mapDefination->insertionOpts.maxDistForInterpolatePoints =
-      m_ui->maxDistForInterpolatePoints->value();
+      static_cast<float>(m_ui->maxDistForInterpolatePoints->value());
   mapDefination->insertionOpts.insertInvalidPoints = m_ui->insertInvalidPoints->isChecked();
 }
